Move TopRightMenu component out of Comment and into own file
